temporary fix to make the art load
turns out the art location for sonic & tails was identified, but NOT where it loads from... Explaining why you only see garbage.
To put it in other words, by fixing the driver, it's no longer aligned to it's original location, thus shifting things around and causing the bug to manifest.